requirements.txt
streamlit==1.52.2
pandas==2.2.3
numpy==2.1.3
plotly==6.5.0
bcrypt==5.0.0
mysql-connector-python==9.1.0
openpyxl==3.1.5
xlrd==2.0.1
python-dotenv==1.0.1
